Attacks
ffgsm_dct and pgd_dct have small changes for solving issues with the image boundary after the idct operation.
stochastic_attack is a random attack. Name is not random due to conflict with random package in python.
slide is a L1 attack.
pgd_fft is a pgd with a Fast Fourier Transform transformation. Analogous to pgd_dct.